Lokalisierungsverhalten verbessert, Init-Code fuer MCU und PC getrennt

+++ mcu/bot-2-sim.c
@@ -70,7 +70,7 @@ void bot_2_sim_inform(void) {
 #ifndef BPS_AVAILABLE
 	command_write(CMD_SENS_LDR, SUB_CMD_NORM , sensLDRL, sensLDRR, 0);
 #else
-	command_write(CMD_SENS_BPS, SUB_CMD_NORM , sensBPS, 0, 0);
+	command_write(CMD_SENS_BPS, SUB_CMD_NORM , (int16_t) sensBPS, 0, 0);
 #endif // BPS_AVAILABLE
 
 #ifdef LED_AVAILABLE
